Refactoring system tests * Re-using launch files in nav2_bringup * Consolidating multi and single robot tests * Removing turtlebot pkgs dependencies
