+++ nav2_bringup/bringup/launch/nav2_multi_tb3_simulation_launch.py
@@ -116,7 +116,7 @@ def generate_launch_description():
         spawn_robots_cmds.append(
             IncludeLaunchDescription(
                 PythonLaunchDescriptionSource(os.path.join(bringup_dir, 'launch',
-                                                           'spawn_robot_launch.py')),
+                                                           'spawn_tb3_launch.py')),
                 launch_arguments={
                                   'x_pose': TextSubstitution(text=str(robot['x_pose'])),
                                   'y_pose': TextSubstitution(text=str(robot['y_pose'])),
