+++ nav2_bringup/spawn_robot/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/spawn_robot
+[install]
+install-scripts=$base/lib/spawn_robot
