--- nav2_bringup/spawn_turtlebot3/package.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>spawn_turtlebot3</name>
-  <version>0.2.4</version>
-  <description>Package for spawning Turtlebot3 into Gazebo</description>
-  <maintainer email="carlos.orduno@intel.com">lkumarbe</maintainer>
-  <maintainer email="lalit.kumar.begani@intel.com">lkumarbe</maintainer>
-  <license>Apache-2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <exec_depend>rclpy</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>python3-pytest</test_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
