--- nav2_bringup/spawn_turtlebot3/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/spawn_turtlebot3
-[install]
-install-scripts=$base/lib/spawn_turtlebot3
