+++ nav2_system_tests/models/turtlebot3_waffle/model.config
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+
+<model>
+  <name>TurtleBot3(Waffle)</name>
+  <version>2.0</version>
+  <sdf version="1.4">model-1_4.sdf</sdf>
+  <sdf version="1.5">model.sdf</sdf>
+
+  <author>
+    <name>Taehun Lim(Darby)</name>
+    <email>thlim@robotis.com</email>
+  </author>
+
+  <description>
+    TurtleBot3 Waffle
+  </description>
+</model>
