+++ nav2_system_tests/src/system/CMakeLists.txt
@@ -38,4 +38,15 @@ ament_add_test(test_dynamic_obstacle
     GAZEBO_MODEL_PATH=${PROJECT_SOURCE_DIR}/models
     BT_NAVIGATOR_XML=navigate_w_replanning_and_recovery.xml
     ASTAR=True
+
+ament_add_test(test_multi_robot
+  GENERATE_RESULT_FOR_RETURN_CODE_ZERO
+  COMMAND "${CMAKE_CURRENT_SOURCE_DIR}/test_multi_robot_launch.py"
+  WORKING_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}"
+  TIMEOUT 120
+  ENV
+    TEST_WORLD=${PROJECT_SOURCE_DIR}/worlds/world_only.model
+    TEST_URDF=${PROJECT_SOURCE_DIR}/urdf/turtlebot3_waffle.urdf
+    TEST_SDF=${PROJECT_SOURCE_DIR}/models/turtlebot3_waffle/model.sdf
+    BT_NAVIGATOR_XML=navigate_w_replanning_and_recovery.xml
 )
